refactor: organize files under views directory
