+++ src/lib/types/faq.ts
@@ -1,0 +1,10 @@
+export interface ContactInfoType {
+  id: string;
+  text: string;
+  content: string;
+}
+
+export interface FAQType {
+  question: string;
+  answer: string;
+}
