--- src/views/FAQPage/types/index.ts
@@ -1,10 +1,0 @@
-export interface ContactInfoType {
-  id: string;
-  text: string;
-  content: string;
-}
-
-export interface FAQType {
-  question: string;
-  answer: string;
-}
