--- src/views/MainPage/lib/index.ts
@@ -1,1 +1,0 @@
-export * from './useHorizontalScroll';
