+++ src/views/ReviewPage/components/index.ts
@@ -1,3 +1,6 @@
-export * from './Description';
-export * from './Reviews';
-export * from './TabBar';
+import Content from './Content';
+import Description from './Description';
+import Reviews from './Reviews';
+import TabBar from './TabBar';
+
+export { Content, Description, Reviews, TabBar };
