+++ src/views/RulesPage/components/CollapseLi/CollapseLi.style.tsx
