+++ src/views/RulesPage/components/CollapseLi/CollapseLi.tsx
