+++ src/views/RulesPage/components/CollapseLi/index.tsx
