+++ src/views/RulesPage/components/RulesHead/RulesHead.tsx
@@ -1,5 +1,4 @@
 import Head from 'next/head';
-import React from 'react';
 
 function RulesHead() {
   return (
