+++ src/views/RulesPage/components/RulesHead/index.tsx
