+++ src/views/RulesPage/components/RulesList/RulesList.style.tsx
