+++ src/views/RulesPage/components/RulesList/index.tsx
