+++ src/views/RulesPage/components/UnderlinedText/UnderlinedText.style.tsx
