+++ src/views/RulesPage/components/UnderlinedText/UnderlinedText.tsx
