+++ src/views/RulesPage/components/UnderlinedText/index.ts
