+++ src/views/RulesPage/components/index.ts
@@ -1,0 +1,6 @@
+import CollapseLi from './CollapseLi/CollapseLi';
+import RulesHead from './RulesHead/RulesHead';
+import RulesList from './RulesList/RulesList';
+import UnderlinedText from './UnderlinedText/UnderlinedText';
+
+export { CollapseLi, RulesHead, RulesList, UnderlinedText };
